nrt: logr: add support for logr
Move to the logr.Logger interface everywhere in the core
source file. This enable named logger, presetting values
for simple and automatic consistency, enables pluggable loggers
and comes for free since we already depend on the logr package
and klog has a native logr integration.

In addition, add minimal support to make it easy to
replace the logr reference, to help integrators of
this code. The default is still (and will still be)
klog for backward compatibility and ecosystem integration.

Signed-off-by: Francesco Romani <fromani@redhat.com>