nrt: logr: always use logging to generate logIDs
reduce the dep to the klog package but, more important,
switch to ha helper which guarantee by design the representation
we need.

Note: a future revision is planned again. If the `klog.KObj`
representation which we were using liberally and interchangeably
actually guarantees the format we need, we can consider
swithcing back (in all the places) to it.

Finally, move all the code to generate logIDs into the
logging package: this include also the time-based generation
used in the resync loop.

Signed-off-by: Francesco Romani <fromani@redhat.com>